astro_project/requirements.txt
Django>=4.2,<5.0
Pillow>=9.0
pyswisseph==2.10.03
